feat: migrate to multimodule, consolidate dependency management and plugin configuration into parent pom.
