+++ .github/workflows/maven-deploy.yml
@@ -1,7 +1,7 @@
 # This workflow will build a package using Maven and then publish it to GitHub packages when a release is created
 # For more information see: https://github.com/actions/setup-java/blob/main/docs/advanced-usage.md#apache-maven-with-a-settings-path
 
-name: Maven Deploy `at_client` to Central Portal
+name: Deploy to Central Portal
 
 on:
   push:
@@ -13,9 +13,6 @@ permissions:  # added using https://github.com/step-security/secure-workflows
 
 jobs:
   deploy:
-    defaults:
-      run:
-        working-directory: at_client
     runs-on: ubuntu-latest
     steps:
     - uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd # v6.0.2
@@ -29,20 +26,10 @@ jobs:
         server-password: "CENTRAL_TOKEN" # env variable for token in deploy
         gpg-private-key: ${{ secrets.GPG_PRIVATE_KEY }} # Value of the GPG private key to import
         gpg-passphrase: "GPG_PASSPHRASE" # env variable for GPG private key passphrase
-
-    - name: Print folder
-      run: ls -al .
-
-    - name: Print m2 settings
-      continue-on-error: true
-      run: cat ~/.m2/settings.xml
-
+    - name: Add virtual env entry to hosts file
+      run: echo "127.0.0.1    vip.ve.atsign.zone" | sudo tee -a /etc/hosts
     - name: Publish to Central
-      run: |
-        mvn \
-        deploy \
-        --batch-mode \
-        --no-transfer-progress \
+      run: mvn deploy --batch-mode --no-transfer-progress
       env:
         CENTRAL_USERNAME: ${{ secrets.CENTRAL_USERNAME }}
         CENTRAL_TOKEN: ${{ secrets.CENTRAL_TOKEN }}
