--- at_client/src/integration-test/java/org/atsign/OnboardingIT.java
@@ -1,22 +1,0 @@
-package org.atsign;
-
-import org.junit.After;
-import org.junit.Before;
-import org.junit.Test;
-
-import static org.junit.Assert.assertTrue;
-
-public class OnboardingIT {
-    @Before
-    public void setUp() {
-    }
-
-    @After
-    public void tearDown() {
-    }
-
-    @Test
-    public void testingIntegrationTests() {
-        assertTrue(true);
-    }
-}
