+++ at_shell/README.md
@@ -1,0 +1,15 @@
+# at_shell
+
+Interactive command line utility which uses the Java at_client library
+
+## Building
+
+```shell
+mvn clean package
+```
+
+## Usage
+
+```shell
+java -jar target/at_shell-0.0.1-SNAPSHOT.jar
+```
