+++ at_shell/demoRepl
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+java -jar target/at_shell-0.0.1-SNAPSHOT.jar org.atsign.client.cli.REPL vip.ve.atsign.zone:64 $1 true
