+++ at_utils/src/main/java/org/atsign/client/util/CameraUtil.java
