+++ examples/README.md
@@ -1,0 +1,1 @@
+Java at_client usage examples
